requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
MouseInfo==0.1.3
mss==9.0.2
numpy @ file:///C:/b/abs_54abayvc9j/croot/numpy_and_numpy_base_1682520598361/work
openai==1.61.0
opencv-python==4.10.0
opencv-python-headless==4.10.0
pillow==10.4.0
PyAutoGUI==0.9.54
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
requests==2.32.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
